Переносит рендеринг снега в воркер
