+++ src/includes/blocks/snow-25.njk
@@ -1,1 +1,1 @@
-<canvas id="snowCanvas"></canvas>
+<canvas role="presentation" id="snowCanvas"></canvas>
